# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ontosyx API",
    "description": "The Semantic Orchestrator — Knowledge Graph Lifecycle Platform.\n\n## Response envelope\n\nEvery successful 2xx response is wrapped in:\n\n```json\n{ \"data\": <T>, \"pagination\": {\"next_cursor\": \"...\"}?, \"meta\": {...}? }\n```\n\nEach handler's `responses(... body = T)` documents the type of `data`, **not** the wire shape. Generated clients should either unwrap `data` in a post-processing step (the approach `web/src/lib/api/client.ts` takes — frontend callers receive `T` directly) or wrap every response type in the envelope shape when consuming the spec from a third-party codegen.\n\nError responses use a separate envelope: `{ \"error\": { \"type\": \"...\", \"message\": \"...\" } }` (see the `ErrorResponse` schema).",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/ontology/export": {
      "post": {
        "tags": [
          "Ontologies"
        ],
        "operationId": "export_ontology",
        "requestBody": {
          "description": "OntologyIR to export",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "InputOntologyDef exchange format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/ontology/export/cypher": {
      "post": {
        "tags": [
          "Ontologies"
        ],
        "operationId": "export_cypher",
        "requestBody": {
          "description": "OntologyIR",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Cypher DDL statements",
            "content": {
              "text/plain": {}
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/ontology/export/graphql": {
      "post": {
        "tags": [
          "Ontologies"
        ],
        "operationId": "export_graphql",
        "requestBody": {
          "description": "OntologyIR",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "GraphQL schema",
            "content": {
              "text/plain": {}
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/ontology/export/mermaid": {
      "post": {
        "tags": [
          "Ontologies"
        ],
        "operationId": "export_mermaid",
        "requestBody": {
          "description": "OntologyIR",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Mermaid ER diagram",
            "content": {
              "text/plain": {}
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/ontology/export/owl": {
      "post": {
        "tags": [
          "Ontologies"
        ],
        "operationId": "export_owl",
        "requestBody": {
          "description": "OntologyIR",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OWL/Turtle ontology",
            "content": {
              "text/plain": {}
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/ontology/export/python": {
      "post": {
        "tags": [
          "Ontologies"
        ],
        "operationId": "export_python",
        "requestBody": {
          "description": "OntologyIR",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Python dataclass definitions",
            "content": {
              "text/plain": {}
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/ontology/export/shacl": {
      "post": {
        "tags": [
          "Ontologies"
        ],
        "operationId": "export_shacl",
        "requestBody": {
          "description": "OntologyIR",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SHACL shapes in Turtle format",
            "content": {
              "text/plain": {}
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/ontology/export/typescript": {
      "post": {
        "tags": [
          "Ontologies"
        ],
        "operationId": "export_typescript",
        "requestBody": {
          "description": "OntologyIR",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "TypeScript interface definitions",
            "content": {
              "text/plain": {}
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/ontology/import/owl": {
      "post": {
        "tags": [
          "Ontologies"
        ],
        "operationId": "import_owl",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OntologyImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Parsed OntologyIR",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Parse or validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/ontology/normalize": {
      "post": {
        "tags": [
          "Ontologies"
        ],
        "operationId": "normalize_ontology",
        "requestBody": {
          "description": "InputOntologyDef to normalize",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Normalized OntologyIR",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "type",
          "message"
        ],
        "properties": {
          "details": {
            "description": "Optional additional details"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "type": {
            "type": "string",
            "description": "Machine-readable error type (e.g., \"not_found\", \"bad_request\")"
          }
        }
      },
...
      "OntologyImportRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string",
            "description": "OWL ontology in Turtle format."
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
...
}